requirements.txt
actionlib==1.12.0
angles==1.9.12
argon2-cffi @ file:///tmp/build/80754af9/argon2-cffi_1613037492802/work
async-generator @ file:///home/ktietz/src/ci/async_generator_1611927993394/work
attrs @ file:///tmp/build/80754af9/attrs_1620827162558/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
bleach @ file:///tmp/build/80754af9/bleach_1612211392645/work
bondpy==1.8.5
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.26
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246439577/work
Click==7.0
cloudpickle==1.2.2
contextlib2==0.6.0.post1
controller-manager==0.18.1
controller-manager-msgs==0.18.1
cv-bridge==1.13.0
cycler==0.10.0
decorator==4.4.1
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distributed-sparse-gp @ file:///home/jamesdi1993/workspace/Distributed-Sparse-GP
dm-sonnet==1.35
dynamic-reconfigure==1.6.3
entrypoints==0.3
ffmpeg==1.4
future==0.18.2
gast==0.2.2
gazebo-plugins==2.8.7
gazebo-ros==2.8.7
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.12
google-pasta==0.1.8
h5py==2.10.0
image-geometry==1.13.0
imageio==2.8.0
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1617877310517/work
interactive-markers==1.11.5
ipykernel @ file:///tmp/build/80754af9/ipykernel_1596206602906/work/dist/ipykernel-5.3.4-py3-none-any.whl
ipython @ file:///tmp/build/80754af9/ipython_1593447367857/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
joblib==0.14.0
joint-state-publisher==1.12.15
jsonschema @ file:///tmp/build/80754af9/jsonschema_1602607155483/work
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-core @ file:///tmp/build/80754af9/jupyter_core_1612213308682/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
kdl-parser-py==1.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.2.0
laser-geometry==1.6.5
Markdown==3.1.1
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1620984851015/work
matplotlib==3.2.2
message-filters==1.14.6
mistune==0.8.4
nb-conda==2.2.1
nb-conda-kernels @ file:///tmp/build/80754af9/nb_conda_kernels_1606775948551/work
nbclient @ file:///tmp/build/80754af9/nbclient_1614364831625/work
nbconvert @ file:///tmp/build/80754af9/nbconvert_1601914804165/work
nbformat @ file:///tmp/build/80754af9/nbformat_1617383369282/work
nest-asyncio @ file:///tmp/build/80754af9/nest-asyncio_1613680548246/work
networkx==2.4
notebook @ file:///tmp/build/80754af9/notebook_1616443456543/work
numpy==1.18.3
opencv-contrib-python==3.4.2.17
packaging @ file:///tmp/build/80754af9/packaging_1611952188834/work
pandas==1.1.5
pandocfilters @ file:///tmp/build/80754af9/pandocfilters_1605120937332/work
parso @ file:///tmp/build/80754af9/parso_1617223946239/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==7.1.2
prometheus-client @ file:///tmp/build/80754af9/prometheus_client_1618088486455/work
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1616415428029/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
pygame==1.9.6
Pygments @ file:///tmp/build/80754af9/pygments_1615143339740/work
pyparsing==2.4.7
pyrsistent @ file:///tmp/build/80754af9/pyrsistent_1600141725711/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
python-qt-binding==0.4.2
pytz==2021.1
-e git+https://github.com/jamesdi1993/cse252c_hw1@ce2974de48b936dfdf1e0bb5f1dd8190cbc46339#egg=pyviso2&subdirectory=pyviso/src
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
qt-dotgraph==0.4.1
qt-gui==0.4.1
qt-gui-cpp==0.4.1
qt-gui-py-common==0.4.1
resource-retriever==1.12.6
rosbag==1.14.6
rosboost-cfg==1.14.9
rosclean==1.14.9
roscreate==1.14.9
rosgraph==1.14.6
roslaunch==1.14.6
roslib==1.14.9
roslint==0.11.2
roslz4==1.14.6
rosmake==1.14.9
rosmaster==1.14.6
rosmsg==1.14.6
rosnode==1.14.6
rosparam==1.14.6
rospy==1.14.6
rosservice==1.14.6
rostest==1.14.6
rostopic==1.14.6
rosunit==1.14.9
roswtf==1.14.6
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.9
rqt-dep==0.4.9
rqt-graph==0.4.11
rqt-gui==0.5.1
rqt-gui-py==0.5.1
rqt-image-view==0.4.14
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.9
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.1
rqt-py-console==0.4.8
rqt-reconfigure==0.5.1
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.9
rqt-robot-steering==0.5.10
rqt-runtime-monitor==0.5.7
rqt-rviz==0.6.0
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.11
rqt-web==0.4.8
rviz==1.13.12
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.11.2
semantic-version==2.8.2
Send2Trash @ file:///tmp/build/80754af9/send2trash_1607525499227/work
sensor-msgs==1.12.7
Shapely==1.7.1
six @ file:///home/linux1/recipes/ci/six_1610970791821/work
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
tabulate==0.8.9
tensorflow-probability==0.8.0
termcolor==1.1.0
terminado==0.9.4
testpath @ file:///home/ktietz/src/ci/testpath_1611930608132/work
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
threadpoolctl==2.1.0
topic-tools==1.14.6
tornado @ file:///tmp/build/80754af9/tornado_1606942266872/work
tqdm==4.37.0
traitlets==4.3.3
typing-extensions @ file:///home/ktietz/src/ci_mi/typing_extensions_1612808209620/work
urdfdom-py==0.4.3
wcwidth @ file:///tmp/build/80754af9/wcwidth_1593447189090/work
webencodings==0.5.1
wrapt==1.11.2
xacro==1.13.5
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
